// wxkanban-agent/core/schemas/lifecycle.d.ts
export declare enum LifecycleStage {
    Design = "Design",
    Implementation = "Implementation",
    QATesting = "QA",
    HumanTesting = "HumanTesting",
    Beta = "Beta",
    Release = "Release"
}
export declare const STAGE_ORDER: readonly LifecycleStage[];
//# sourceMappingURL=lifecycle.d.ts.map